Move ui::Quad to its own header.
- Use `RoundedQuad`.
