+++ scene/header_list.cmake
@@ -8,6 +8,7 @@ set(header_list
   ${prefix}/scene/ui/element.hpp
   ${prefix}/scene/ui/input_text.hpp
   ${prefix}/scene/ui/primitive_renderer.hpp
+  ${prefix}/scene/ui/quad.hpp
   ${prefix}/scene/ui/rect_transform.hpp
   ${prefix}/scene/ui/renderable.hpp
   ${prefix}/scene/ui/text.hpp
