+++ scene/src/ui/CMakeLists.txt
@@ -2,6 +2,7 @@ target_sources(${PROJECT_NAME} PRIVATE
   element.cpp
   input_text.cpp
   primitive_renderer.cpp
+  quad.cpp
   rect_transform.cpp
   renderable.cpp
   text.cpp
