+++ scene/src/ui/primitive_renderer.cpp
@@ -8,12 +8,4 @@ auto PrimitiveRenderer::render(std::vector<graphics::RenderObject>& out) const -
 
 	render_to(out, &material, &primitive);
 }
-
-auto Quad::tick(Duration dt) -> void {
-	if (!is_active()) { return; }
-
-	PrimitiveRenderer::tick(dt);
-
-	set_quad(graphics::Quad{.size = transform.extent});
-}
 } // namespace le::ui
